requirements.txt
altgraph==0.17
aniso8601==8.1.1
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5.4
beautifulsoup4==4.9.1
botocore==1.20.50
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
click==7.1.2
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.2
django-debug-toolbar==2.2
django-excel==0.0.10
django-heroku==0.3.1
django-s3direct==1.1.8
django-storages==1.11.1
djangorestframework==3.12.2
easygui==0.98.1
Eel==0.13.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
google-search==1.0.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.5.0
install==1.3.3
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
lml==0.0.9
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
openpyxl==3.0.3
package-name==0.1
parso==0.6.2
pefile==2019.4.18
Pillow==8.0.1
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyexcel==0.6.4
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyflakes==2.1.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pylint==2.4.4
PyMySQL==0.10.0
PyPDF2==1.26.0
PyQt5==5.14.1
PyQt5-sip==12.7.1
pytesseract==0.3.6
python-dateutil==2.8.1
python-decouple==3.4
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
pytz==2019.3
pywin32-ctypes==0.2.0
r-django-jet==2.0.9
reportlab==3.5.59
requests==2.25.1
s3transfer==0.3.6
six==1.13.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.23
sqlparse==0.3.0
texttable==1.6.2
toml==0.10.1
typed-ast==1.4.0
urllib3==1.26.4
virtualenv==16.7.9
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
wrapt==1.11.2
zipp==3.0.0
zope.event==4.4
zope.interface==5.1.0
